+++ src/panels/FocusedModePanel.tsx
@@ -66,7 +66,6 @@ type Props = Pick<
   | 'onApplyTextNodeMarkdown'
   | 'onSetTextDocumentContent'
   | 'onSetTextDocumentBlockGap'
-  | 'onResetCodeBlockStyle'
   | 'onEnterFocusedMode'
   | 'onActivateRichEdit'
   | 'onOpenManageFonts'